language/wino_dict/requirements.txt
absl-py~=1.1.0
nltk~=3.7
spacy~=3.3.1
tensorflow~=2.8.2
tfds_nightly~=4.6.0
